Refactoring

+++ datasets/seq_eurosat.py
@@ -45,7 +45,7 @@ class MyEuroSAT(Dataset):
         self.data = np.array(
             [Image.open(self.root + "/EuroSAT_RGB/" + img).convert("RGB") for img in self.data_split[0].values]
         )
-        self.targets = self.data_split[1].values
+        self.targets = self.data_split[1].values.astype(np.int64)
 
     def __len__(self):
         return len(self.targets)
@@ -72,17 +72,17 @@ class SequentialEuroSAT(BaseDataset):
     MEAN, STD = [0.48145466, 0.4578275, 0.40821073], [0.26862954, 0.26130258, 0.27577711]
     normalize = transforms.Normalize(mean=MEAN, std=STD)
     TRAIN_TRANSFORM = transforms.Compose(
-        # transforms.RandomResizedCrop(224, scale=(0.08, 1.0), interpolation=InterpolationMode.BICUBIC),  # from https://github.dev/KaiyangZhou/Dassl.pytorch defaults
-        # transforms.RandomHorizontalFlip(),
         [
+            # transforms.RandomResizedCrop(224, scale=(0.08, 1.0), interpolation=InterpolationMode.BICUBIC),  # from https://github.dev/KaiyangZhou/Dassl.pytorch defaults
+            # transforms.RandomHorizontalFlip(),
             transforms.ToTensor(),
             # normalize,
         ]
     )
     TEST_TRANSFORM = transforms.Compose(
-        # transforms.Resize(224, interpolation=3),  # bicubic
-        # transforms.CenterCrop(224),
         [
+            # transforms.Resize(224, interpolation=3),  # bicubic
+            # transforms.CenterCrop(224),
             transforms.ToTensor(),
             # normalize,
         ],
